Move to case defaults library (#5)
* case-defaults

* Move case defaults to case-defaults library.

* Delete CustomFormats.java

* Removed TitleParser.
