--- src/main/java/no/fint/p360/CustomFormats.java
@@ -1,17 +1,0 @@
-package no.fint.p360;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-import java.util.Map;
-
-@Data
-@Component
-@EnableConfigurationProperties
-@ConfigurationProperties("fint.p360.format")
-public class CustomFormats {
-    private Map<String, String> title;
-    private Map<String, Map<String, String>> field;
-}
